requirements.txt
blinker
click
colorama
dnspython
Flask
itsdangerous
Jinja2
MarkupSafe
pdfcrowd
pymongo
Werkzeug
python-dotenv